Generate the token when the user login with user login button and store into the firebase and if the button is already exsist the update the token with the new token.
Solve some Basic errors like in the userchat.dart (the error is not displaying the proper name of the students).

Adding the Download button in the app bar so that they can navigate to downloader.dart where there is a two screen pdfDownloads.dart and videoDownloads.dart where shows the downloaded data to the user.